+++ project2/filemanager/file.h
@@ -46,7 +46,7 @@ void _flush_header();
  * @param   path    Database file path.
  * @return          ID of the opened database file.
  */
-int64_t file_open_database_file(const char* path);
+int file_open_database_file(const char* path);
 
 /**
  * @brief   Allocate an on-disk page from the free page list
